Blog-index "_entry" partial now calls 'PostDecorator#build_body' method.
Safe in the knowledge that the passed-in `entry` *is* in fact decorated,
courtesy of the last commit.

Also updated a feature spec to prove that the rendered page has the
markup for the image, with the post body as its caption.

RSpec: 215 examples, 0 failures
Script: 0 tests: 0 failures, 0 errors, 0 pending, 0 omitted
RuboCop: 63 files inspected, no offenses detected
RuboCop global overrides in `.rubocop.yml`: 0
RuboCop local overrides in our source/spec files: 0