New ApplicationHelper function `#build_menu_for` and specs.
`#build_menu_for` builds the menu for either the app-layout nav bar or
sidebar. The next commit will include a working sidebar implementation
to match this commit's navbar one, as well as calls from the application
layout and navbar partial.

Once we have this (with the current everything-at-once menu), we can
have it build *just* the menu for the Guest User or the logged-in user.

RSpec: 330 examples, 0 failures
RuboCop: 86 files inspected, no offenses detected