Final bit of Step 2 of the 'latest...strategy' from Issue #34.
See https://github.com/jdickey/new_poc/issues/34#issuecomment-49047556
for the spec; basically a '"traditional" implementation of Pundit as a
first step to what's coming, and to prove out what to me is new tech.'
This was a *large* brick exiting the orifice laterally, and cleanup
touched all the other authentically resource-based controllers that had
already been adapted. The kicker, though, was getting login and logout
to behave enough like a RESTful resource that Pundit, and ActiveModel,
could deal with it. Login (#create) and logout (#destroy) were
predictably gnarly (in the "earn"-while-you-learn sense); what broke the
feature specs at the very end (and thus the app in a browser) was the
fact that there needed to be a #new action, properly authorised, in
order to get to #create anything. Which makes perfect sense -- in
hindsight. :anguished: Oh, well; the stated objective WAS to learn the
new-to-me tech.

RSpec: 389 examples, 0 failures
RuboCop: 111 files inspected, no offenses detected