A challenge and a conundrum, moving Pundit into DSOs.
This should be fully documented in the first ~60 lines of
SessionsController; the reader is invited to peruse the (uncommented)
code in the controller's #new method and compare it to the #execute
method of the DSO, while keeping in mind the linked code submitted as a
PR for Pundit (adding PunditAuthorizer).

On a related track, if anyone can explain to me why the Obvious Code for
what is now lines 26-27 does *not* Just Work, and even better, how to
have a less egregious hack there, I would greatly appreciate it.

RSpec: 389 examples, 0 failures
RuboCop: 110 files inspected, no offenses detected