Backtracking a bit.
This *should* be where we close the issue for this branch. Instead,
we've a major WTF that's had me stumped for far too long.  Look at the
code for `app/helpers/users_helper/profile_bio_panel_builder.rb`, in
the `#to_html` method. Compare that to, say,
`profile_article_list_builder` in the same directory. Note the
resemblance that the latter has to the commented-out code in the
former. Why commented out and replaced with raw strings, which are a
known code stench? Because, despite the parameters to Rails
`#content_tag` helper, it insists on escaping the HTML in the
`@user_profile` string when viewed in a browser. Specs currently pass,
of course. Any resolution gratefully appreciated.

So, for the moment, I plan to merge this branch back into master and go
fry the bigger fish waiting in my pond, or cesspool as the case may be.

----

For what it's worth,

RSpec: 453 examples, 0 failures
RuboCop: 126 files inspected, no offenses detected