Verifying understanding of RedCarpet options pt 12
Documented current behaviour of/defaults for
`: highlight ` option.

Note that this has a not-entirely-intuitive syntax that makes it
trivially easy to break the markup such that the feature appears not to
work. Omitting *any* of the following will break parsing:

* the ^ before the footnote number in either the reference or
definition;
* the colon after the closing bracket in the footnote definition (see
line 573); or
* the newline at the end of the footnote definition;
* *multiple* blank lines (newlines) before the footnote definition.

Anybody that wants to use this *must* supply a preview; it's simply too
easy to screw it up.

RSpec: 492 examples, 0 failures
RuboCop: 127 files inspected, no offenses detected