New UserUpdater DSO for UsersController#update.
The new DSO, given a User entity and a hash of all-optional new
attributes to set on it, modifies the entity and returns a copy of the
(updated) hash of modifiable attributes from `#execute`. Why not the
entity itself? Rails (4.x, anyway) really wants us to call
`#update_attributes` to do that; `#save` is to be used for a new
record. Oh, well.

Next up: adding an edit button to the profile-view page and finally
closing Issue #61.

Other changes:
* The User CCO now knows about the `:slug` attribute;
* Minor version updates in Gemfile.lock;
* UsersControllerSpec has specs for `#update` fully(?) fleshed out;
* User CCO spec knows about the slug; creates instead of building
implementation objects to trigger slug creation.

RSpec: 537 examples, 0 failures
RuboCop: 134 files inspected, no offenses detected