[Fix #82] Timestamps in views use HH:MM format.
Over the past few commits, this has touched several files, and a
surprising number of specs. Why? Because too many specs know what the
format of the timestamps they're looking for is supposed to be. Enter
the `DecoratorShared#timestamp_format` and
`FeatureSpecTimestampHelper#to_timestamp_s` methods. The specs for the
former are now the *only* specs that should know about the format;
everything else defers to it or, for feature specs, to the latter.

RSpec: 657 examples, 0 failures
RuboCop: 151 files inspected, no offenses detected