Sometimes you want to validate *just* the object.
PostUpdateValidator now treats the `data` parameter to #initialize as an
optional parameter, defaulting to an empty hash. This will have the
effect of validating the current `post` (whether entity or ActiveRecord
model).

So why do we still have validation code in the model? That's not the
only, or even the most pressing, question at the moment.

RSpec: 697 examples, 0 failures (but see Issue #87)
RuboCop: 153 files inspected, no offenses detected