`Blog` no longer dependent on `BlogData`.
The `BlogData` implementation model never had much to do besides supply
the one-and-only title and subtitle; we're now injecting that into the
`Blog` entity, with defaults so nothing else breaks. We should be able
to now completely remove the `BlogData` model, related code such as
it's CCO and BLO, and the database table itself.

RSpec: 688 examples, 0 failures (but see Issue #87)
RuboCop: 152 files inspected, no offenses detected