Fixing typo in updated Blog entity.
We'd previously forwarded attribute accessors for `:title`, `:subtitle`
and `:entries` to a `BlogData` instance, which is no longer used here.
This was the easy part of "getting rid of" `BlogData`. More
complicated, as noted earlier, is going to be moving the
`BlogDataDecorator#summarise` method, called by the blog index view.
Should we just transplant that method to the`BlogHelper` module? Should
we break down and add "real" pagination?

RSpec: 691 examples, 0 failures (but see Issue #87)
RuboCop: 154 files inspected, no offenses detected