Done with Issue 100 Phase 1.
Adding a Post to a Blog is no longer synonymous with "publishing" it;
publishing a post causes it to be added to its associated blog if it
hasn't yet been, but that is now decoupled from the publication state.

Also some cleanup on the PostCCO spec, now that we remember what the
correct matcher is. :anguished:

RSpec: 711 examples, 0 failures (but see Issue #87)
RuboCop: 150 files inspected, no offenses detected