PostCCO#from_entity now sets post_status properly.
Remember that `:post_status` is the synthetic attribute we added to the
`PostData` model in the previous commit. This commit updates the
`CCO::PostCCO` cross-layer conversion object to initialise the
attribute correctly based on the `pubdate` attribute of the entity
being converted. (There is presently no need for a status-type field at
the *entity* level; this is being added solely for the
implementation-level UI.)

RSpec: 715 examples, 0 failures (but see Issue #87)
RuboCop: 150 files inspected, no offenses detected