Added PostsHelper#status_select_options method.
This builds the options for a `select` helper (using the
`#options_for_select` form-options helper), using the `post_status`
attribute added previously, and defaulting to 'draft' if that attribute
is somehow not set. This will allow us to add the post-status UI to the
create-post form.

RSpec: 723 examples, 0 failures (but see Issue #87)
RuboCop: 150 files inspected, no offenses detected