PostData#post_status now defaults to 'draft'.
Instead of a simple `attr_accessor` for `post_status`, we have an
`attr_writer` and a custom reader method that sets the return value to
`public` if the `pubdate` field is set. We've also added a
"post_status" field to the `_post_edit_fields` partial. All's well by
the existing specs…

RSpec: 723 examples, 0 failures (but see Issue #87)
RuboCop: 150 files inspected, no offenses detected