Starting replacement Post CCO and specs.
What was the problem with the old one? Posts commingle the concepts of
publication, persistence and attachment to a Blog, courtesy of the
original Book-driven implementation. We've got to prise those apart
from each other, and the Superglue has resisted the first half-dozen
attempts not pushed here. Onward!

RSpec: 734 examples, 0 failures (but see Issue #87)
RuboCop: 153 files inspected, no offenses detected