Alternate fix using existing Blog controller.
The whole point of Use Case 3 on Issue #52 is to have draft posts
visible to the author, but only to the author. Technically, this commit
satisfies that requirement.

This is FUGLY code that should NEVER see the light of day beyond barest
proof-of-concept. The view helper for one controller/view salmagundi
(BlogHelper) still knows much too much too much about the workings of
the PostData model.  Issue #109 will solve this, and should be
completed.

But it's nice to see drafts listed when the author is logged in, then
log out and see them disappear from the index. Ah well.

RSpec: 841 examples, 0 failures (but see Issue #87)
RuboCop: 161 files inspected, no offenses detected