[Fix #87] Correct spec count now being reported.
*That* long saga is now complete; simply moving the specs for "support"
classes out of `spec/support` (and into `spec/support_specs`) and
patching them appropriately let RSpec remember how to count properly
again. The Rakefile needed to have the new path added to the list of
places Rubocop looks for files to inspect, but it took longer to
describe the change than to make it.

Whodathunkit?!?