First Action class and spec.
Omitting the time used to fix a typo-level logic error, total time
elapsed to write code+specs was well under an hour. (Don't ask; just
beware copy-and-paste errors.)

`Actions::NewSession` is the first of our new, lightweight Wisper-based
"action" classes; it broadcasts either `:success` or `:failure` with a
payload of a `StoreResult` instance, that tells the receiver (if any)
everything it needs to know about the Guest User entity (for success)
or the "already logged in as <whoever>" message (for failure).

What happens after that? This puppy *doesn't care*, and that's the
whole beauty of it.

RSpec: 911 examples, 0 failures, 136 pending
RuboCop: 232 files inspected, no offenses detected