Wired Actions::ShowUser into UsersController#show.
The `Actions::ShowUser` action is oblivious to the current-user status;
the *Web presentation* corresponding to this action is supposed to
display an "Edit Profile" button when a user is viewing his own
profile. That's *beyond the scope* of this Action, which simply grabs
the `UserEntity` that will be used for display.

RSpec: 1016 examples, 0 failures, 103 pending
RuboCop: 248 files inspected, no offenses detected