CodeShip *should* be happy now.
We've been a *baaaaaaad* boy. :disappointed:

On a purely local system, we got away with comparing records for
equality when a) those records had timestamp fields that b) were
generated more than two steps apart. Put that on a slow server, and
hilarity ensues. (We don't really know if CIrcleCI was faster because
we didn't get that far yet, but impressions are strange things.)  That
should be fixed now.

On the other hand, most of these are odious, tech-debt-inducing hacks
crying out for better solutions. We'll get there.

RSpec: 772 examples, 0 failures
RuboCop: 174 files inspected, no offenses detected