Replacing RedCarpet w/HTML::Pipeline for parsing.
HTML::Pipeline *in its README* talks about how to parse fenced code
blocks, at-mentions, sanitisation, and all sorts of other Good and
Useful Stuff that we'd have to reinvent adding on top of RedCarpet.

First up: an exploration spec to fully kick the tires and document the
API features we (will) use, in such a way that future breaking API
changes break both our main-app code *and* the exploration spec.
(Because if only our main-app code breaks, it *could* be a bug in *our*
code…).

This commit is the first "tire kicked" in the exploration spec.

RSpec: 691 examples, 0 failures
RuboCop: 179 files inspected, no offenses detected