MelddRepository::StoreResult now up. + WARNING.
Duplicating the soon-to-be-removed `StoreResult` in the main app.

IMPORTANT NOTE: There is a Heisenbug that causes the spec at
spec/controllers/users_controller_spec.rb line 54 to fail
intermittently and unpredictably when and (so far) *only* when the
entire spec suite is run. However, running the suite again with the
same seed as the failure (which should run the specs in the same
otherwise-randomised order) has yet to repeat the failure. This has
been observable for at least the previous two commits and may in fact
go back farther. Resources have not yet been expended in further
exploration.

RSpec: 664+3 examples, 2 failures
RuboCop: 186 files inspected, no offenses detected