Streamlining boilerplate controller actions.
`ApplicationController.def_action` is a new method that, given a symbol
identifying the action method (`:edit`, `:update`, etc) to create and a
block (executed in the instance's context) which provides
initialisation parameters for the Action class, defines the method as
we had been doing it heretofore. This just shoves the namespaces and
Wisper plumbing out of the way.

RSpec: 684 examples, 0 failures, 1 pending
RuboCop: 223 files inspected, no offenses detected