User entity class knowledge now fully contained.
The only place that maps the concept of a "User entity" to a
`Newpoc::Entity::User` instance is now `PostFactory.entity_class`.
Which is as it should have been all along. (The fact that *17 files*
had to change to accomplish this is mortifying.)

RSpec: 715 examples, 0 failures, 1 pending
RuboCop: 234 files inspected, no offenses detected