Added contracts to `NameValidator`.
The equivocation on the parameter for `#initialize` is to handle one
test spec in the spec for `Entity::User` where a new entity has a value
of `nil` for (nearly) all fields, and should simply be regarded as
invalid.

RSpec: 1005 examples, 0 failures, 1 pending
RuboCop: 291 files inspected, no offenses detected
Coverage: 2797 / 2816 LOC (99.33%) covered.